feat(security): token validation added to all endpoints and documented

+++ transactions/views.py
@@ -1,7 +1,10 @@
+from drf_yasg import openapi
 from drf_yasg.utils import swagger_auto_schema
 from rest_framework import viewsets, status
 from rest_framework.response import Response
 
+from commons.jwt_utils import JWTUtils
+from commons.permissions import Permissions
 from transactions.models import Transaction
 from transactions.serializers import TransactionRequestSerializer, TransactionDataSerializer
 from transactions.services import TransactionService
@@ -13,9 +16,23 @@ class TransactionViewSet(viewsets.ViewSet):
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
         self.transaction_service = TransactionService()
+        self.forbidden_response = Response({
+            "message": "You don't have permissions to perform this action.",
+        }, status=status.HTTP_403_FORBIDDEN)
 
-    @swagger_auto_schema(request_body=TransactionRequestSerializer, responses={201: TransactionDataSerializer()})
+    header_param = openapi.Parameter('authorization', openapi.IN_HEADER, description="authorization token header param",
+                                     type=openapi.IN_HEADER)
+
+    @swagger_auto_schema(request_body=TransactionRequestSerializer, responses={201: TransactionDataSerializer()},
+                         manual_parameters=[header_param])
     def create(self, request):
+        if 'authorization' not in request.headers:
+            return self.forbidden_response
+
+        token_info = JWTUtils.decode(request.headers['authorization'])
+        if Permissions.CREATE_TRANSACTION not in token_info['permissions']:
+            return self.forbidden_response
+
         transaction_request_serializer: TransactionRequestSerializer = TransactionRequestSerializer(data=request.data)
         if transaction_request_serializer.is_valid(raise_exception=True):
             transaction, products_per_transaction = transaction_request_serializer.create(transaction_request_serializer.data)
@@ -24,8 +41,16 @@ class TransactionViewSet(viewsets.ViewSet):
             return Response(transaction_serializer.data, status=status.HTTP_201_CREATED)
 
     @swagger_auto_schema(responses={200: TransactionDataSerializer(),
-                                    404: "{'error': 'Transaction not found'}"})
+                                    404: "{'error': 'Transaction not found'}"},
+                         manual_parameters=[header_param])
     def retrieve(self, request, pk=None):
+        if 'authorization' not in request.headers:
+            return self.forbidden_response
+
+        token_info = JWTUtils.decode(request.headers['authorization'])
+        if Permissions.VIEW_TRANSACTION not in token_info['permissions']:
+            return self.forbidden_response
+
         try:
             return Response(self.transaction_service.get_transactions_by_id(pk).to_dict(), status=status.HTTP_200_OK)
         except Transaction.DoesNotExist:
@@ -33,8 +58,15 @@ class TransactionViewSet(viewsets.ViewSet):
                 "error": "Transaction not found",
             },status=status.HTTP_404_NOT_FOUND)
 
-    @swagger_auto_schema(responses={200: TransactionDataSerializer(many=True)})
+    @swagger_auto_schema(responses={200: TransactionDataSerializer(many=True)}, manual_parameters=[header_param])
     def list(self, request):
+        if 'authorization' not in request.headers:
+            return self.forbidden_response
+
+        token_info = JWTUtils.decode(request.headers['authorization'])
+        if Permissions.VIEW_TRANSACTION not in token_info['permissions']:
+            return self.forbidden_response
+
         transactions = self.transaction_service.get_all_transactions()
         response = []
         for transaction in transactions:
@@ -42,8 +74,16 @@ class TransactionViewSet(viewsets.ViewSet):
 
         return Response(response, status=status.HTTP_200_OK)
 
-    @swagger_auto_schema(request_body=TransactionRequestSerializer, responses={200: TransactionDataSerializer()})
+    @swagger_auto_schema(request_body=TransactionRequestSerializer, responses={200: TransactionDataSerializer()},
+                         manual_parameters=[header_param])
     def update(self, request, pk=None):
+        if 'authorization' not in request.headers:
+            return self.forbidden_response
+
+        token_info = JWTUtils.decode(request.headers['authorization'])
+        if Permissions.UPDATE_TRANSACTION not in token_info['permissions']:
+            return self.forbidden_response
+
         transaction_request_serializer: TransactionRequestSerializer = TransactionRequestSerializer(data=request.data)
         if transaction_request_serializer.is_valid(raise_exception=True):
             transaction, _ = transaction_request_serializer.create(transaction_request_serializer.data)
@@ -52,8 +92,16 @@ class TransactionViewSet(viewsets.ViewSet):
             return Response(transaction_saved.to_dict(), status=status.HTTP_200_OK)
 
     @swagger_auto_schema(responses={200: "'message': 'This Transaction has been deleted successfully'",
-                                    404: "{'error': 'Transaction not found'}"})
+                                    404: "{'error': 'Transaction not found'}"},
+                         manual_parameters=[header_param])
     def destroy(self, request, pk=None):
+        if 'authorization' not in request.headers:
+            return self.forbidden_response
+
+        token_info = JWTUtils.decode(request.headers['authorization'])
+        if Permissions.DELETE_TRANSACTION not in token_info['permissions']:
+            return self.forbidden_response
+
         try:
             self.transaction_service.delete_transaction(pk)
             return Response({
